[RDSBUG-558][RDSBUG-568] Rework firmware versioning
This commit reworks the firmware versioning:

1. For backwards compatibility (we noticed a lot of internal tools still
depend on this) the firmware version is now again also saved to
/etc/version in addition to the new location in /etc/os-release. This
is not ideal, as /etc/version originally is a file generated deep
within the yocto build process (which is why the previous versioning
broke in the first place during a yocto update). As such, we cannot
guarantee, that this won't break again in the future. We therefore
recommend using the FIRMWARE_VERSION environment variable from
/etc/os-release instead.

2. In order to automate the flash verification procedure, the firmware
versioning schema within the OS, as well as the update_files are (mostly)
identical (the update_files versioning is additionally prepended with
"firmware-" or "bootloader-", depending on the type of update file).

3. The update_files meta.yml files as well as /etc/os-release
now also contain the "upstream" u-boot version. As such, we can prevent
unnecessary reflashes of the bootloader, when there were no changes to
the bootloader between two (or more) releases.

4. The version string does not contain a timestamp anymore (this was
confusing to some non-techies). We might have to re-add these, if it
turns out, that developers actually need this.

5. Fix a regression introduced in [https://github.com/iris-GmbH/meta-iris-base/commit/8ffcdd92bda24a7ad508e4b37047af1a2e6ade51]
which caused the irma-6-firmware-zip.bbclass to fail at cleaning
previous builds, if the version number changed (the softlink may NOT
contain the version number).

6. irma-six-* has now been changed to irma6-*